Add recombination as a rephasing function.
The current target and best are considered the parents. If the current target phase is best, then this does nothing. Otherwise, it is a random uniform crossover between the two parents on variables that differ.

Something like Partition Crossover would be better, but this requires a dynamic VIG and is probably too expensive to warrant a prototype.